+++ airs/templates/airs/air_detail.html
@@ -128,17 +128,4 @@
 
 {% include 'base/separator.html' %}
 
-
-<br><br><br><br><br><br><br><br><br><br>
-
-{% if error_message %}<p><strong>{{ error_message }}</strong></p>{% endif %}
-
-<form action="{% url 'airs:vote' air.id %}" method="post">{% csrf_token %}
-  {% for nanitozo in air.nanitozo_set.all %}
-  <input type="radio" name="nanitozo" id="nanitozo{{ forloop.counter }}" value="{{ nanitozo.id }}">
-  <label for="nanitozo{{ forloop.counter }}" class="uk-text-break">{{ nanitozo.comment }}</label><br>
-  {% endfor %}
-  <input type="submit" value="Vote" class="uk-button uk-button-default">
-</form>
-
 {% endblock content %}
